feat: replace mockprofile profile-data

+++ src/main/java/org/nfactorial/newsfeed/domain/post/service/PostService.java
@@ -2,16 +2,19 @@ package org.nfactorial.newsfeed.domain.post.service;
 
 import org.nfactorial.newsfeed.common.code.ErrorCode;
 import org.nfactorial.newsfeed.common.exception.BusinessException;
+import org.nfactorial.newsfeed.common.security.AuthProfileDto;
 import org.nfactorial.newsfeed.domain.post.dto.request.PostCreateRequest;
 import org.nfactorial.newsfeed.domain.post.dto.request.PostUpdateRequest;
 import org.nfactorial.newsfeed.domain.post.dto.response.PostCreateResponse;
 import org.nfactorial.newsfeed.domain.post.dto.response.PostGetOneResponse;
 import org.nfactorial.newsfeed.domain.post.dto.response.PostUpdateResponse;
 import org.nfactorial.newsfeed.domain.post.entity.Post;
-import org.nfactorial.newsfeed.domain.post.mock.MockAuthProfileDto;
 import org.nfactorial.newsfeed.domain.post.repository.PostRepository;
+import org.nfactorial.newsfeed.domain.profile.entity.Profile;
+import org.nfactorial.newsfeed.domain.profile.repository.ProfileRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.ObjectUtils;
 
 import lombok.RequiredArgsConstructor;
 
@@ -20,26 +23,28 @@ import lombok.RequiredArgsConstructor;
 public class PostService implements PostServiceApi {
 
 	private final PostRepository postRepository;
+	private final ProfileRepository profileRepository;
 
-	//TODO 프로필 받고 변경
 	@Transactional
-	public PostCreateResponse save(PostCreateRequest request, MockAuthProfileDto currentUserProfile) {
+	public PostCreateResponse save(PostCreateRequest request, AuthProfileDto currentUserProfile) {
+		long profileId = currentUserProfile.profileId();
+		Profile foundProfile = profileRepository.findById(profileId)
+			.orElseThrow(() -> new BusinessException(ErrorCode.PROFILE_NOT_FOUND));
 
-		Post savedPost = postRepository.save(Post.of(request, currentUserProfile));
+		Post savedPost = postRepository.save(Post.of(request, foundProfile));
 
 		return PostCreateResponse.of(savedPost);
 	}
 
-	//TODO 프로필 받고 변경
 	@Transactional
 	public PostUpdateResponse update(Long postId, PostUpdateRequest request,
-		MockAuthProfileDto currentUserProfile) {
+		AuthProfileDto currentUserProfile) {
 
 		Post foundPost = getPostById(postId);
 
-		// if (!ObjectUtils.nullSafeEquals(foundPost.getProfile().getId(), currentUserProfile.getId())) {
-		// 	throw new BusinessException(ErrorCode.POST_ACCESS_DENIED);
-		// }
+		if (!ObjectUtils.nullSafeEquals(foundPost.getProfile().getId(), currentUserProfile.profileId())) {
+			throw new BusinessException(ErrorCode.POST_ACCESS_DENIED);
+		}
 
 		foundPost.updateContent(request);
 
@@ -58,13 +63,13 @@ public class PostService implements PostServiceApi {
 	}
 
 	@Transactional
-	public void deleteById(Long postId, MockAuthProfileDto currentUserProfile) {
+	public void deleteById(Long postId, AuthProfileDto currentUserProfile) {
 
 		Post foundPost = getPostById(postId);
 
-		// if (!ObjectUtils.nullSafeEquals(foundPost.getProfile().getId(), currentUserProfile.getId())) {
-		// 	throw new BusinessException(ErrorCode.POST_ACCESS_DENIED);
-		// }
+		if (!ObjectUtils.nullSafeEquals(foundPost.getProfile().getId(), currentUserProfile.profileId())) {
+			throw new BusinessException(ErrorCode.POST_ACCESS_DENIED);
+		}
 
 		postRepository.delete(foundPost);
 	}
